pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["universal_policies"]

[project]
name = "universal_policies"
version = "0.1.0"
dependencies = [
"aiohttp==3.8.6",
"aiosignal==1.3.1",
"appdirs==1.4.4",
"async-timeout==4.0.3",
"attrs==23.1.0",
"av==10.0.0",
"black==23.9.1",
"blosc==1.11.1",
"certifi==2023.7.22",
"charset-normalizer==3.3.0",
"click==8.1.7",
"cloudpickle==2.2.1",
"contourpy==1.1.1",
"cycler==0.12.1",
"decorator==4.4.2",
"docker-pycreds==0.4.0",
"einops==0.7.0",
"Farama-Notifications==0.0.4",
"filelock==3.12.4",
"fonttools==4.43.1",
"frozenlist==1.4.0",
"fsspec==2023.9.2",
"gitdb==4.0.11",
"GitPython==3.1.40",
"gymnasium==0.29.1",
"idna==3.4",
"imageio==2.32.0",
"imageio-ffmpeg==0.4.9",
"Jinja2==3.1.2",
"kiwisolver==1.4.5",
"lightning==2.1.0",
"lightning-utilities==0.9.0",
"MarkupSafe==2.1.3",
"matplotlib==3.8.0",
"moviepy==1.0.3",
"mpmath==1.3.0",
"multidict==6.0.4",
"mypy-extensions==1.0.0",
"networkx==3.1",
"numpy==1.26.0",
"nvidia-cublas-cu12==12.1.3.1",
"nvidia-cuda-cupti-cu12==12.1.105",
"nvidia-cuda-nvrtc-cu12==12.1.105",
"nvidia-cuda-runtime-cu12==12.1.105",
"nvidia-cudnn-cu12==8.9.2.26",
"nvidia-cufft-cu12==11.0.2.54",
"nvidia-curand-cu12==10.3.2.106",
"nvidia-cusolver-cu12==11.4.5.107",
"nvidia-cusparse-cu12==12.1.0.106",
"nvidia-nccl-cu12==2.18.1",
"nvidia-nvjitlink-cu12==12.2.140",
"nvidia-nvtx-cu12==12.1.105",
"packaging==23.2",
"pathspec==0.11.2",
"pathtools==0.1.2",
"Pillow==10.0.1",
"platformdirs==3.11.0",
"proglog==0.1.10",
"protobuf==4.24.4",
"psutil==5.9.6",
"pygame==2.5.2",
"pyparsing==3.1.1",
"python-dateutil==2.8.2",
"pytorch-lightning==2.1.0",
"PyYAML==6.0.1",
"requests==2.31.0",
"rotary-embedding-torch==0.3.5",
"sentry-sdk==1.32.0",
"setproctitle==1.3.3",
"six==1.16.0",
"smmap==5.0.1",
"sympy==1.12",
"tomli==2.0.1",
"torch==2.0.0",
"torchmetrics==1.2.0",
"torchvision==0.15.1",
"tqdm==4.66.1",
"triton==2.0.0",
"typing_extensions==4.8.0",
"urllib3==2.0.7",
"wandb==0.15.12",
"yarl==1.9.2"
]
requires-python = ">=3.11"
[project.optional-dependencies]
dev = [
    "pre_commit==4.1.0"
]